--- Services/Auth/Auth.Application/Commands/CreateUserCommand.cs
@@ -1,20 +1,0 @@
-﻿using User.Application.Responses;
-using MediatR;
-
-namespace User.Application.Commands
-{
-    public class CreateUserCommand : IRequest<AuthResponse>
-    {
-        public string NickName { get; set; } = null!;
-        public string Email { get; set; } = null!;
-        public string Password { get; set; } = null!;
-
-        public CreateUserCommand(string nickName, string email, string password)
-        {
-            NickName = nickName;
-            Email = email;
-            Password = password;
-        }
-
-    }
-}
