+++ Services/Post/Post.API/Controllers/PostsController.cs
@@ -1,6 +1,5 @@
 ﻿using MediatR;
 using Microsoft.AspNetCore.Mvc;
-using System.Net;
 using Post.Application.Commands;
 using Post.Application.Queries;
 using Post.Application.Responses;
