+++ Services/User/User.API/Controllers/UsersController.cs
@@ -1,9 +1,9 @@
-﻿using User.Application.Commands;
-using User.Application.Queries;
-using User.Application.Responses;
+﻿using User.Application.Responses;
 using MediatR;
 using Microsoft.AspNetCore.Mvc;
 using User.Domain.DTO;
+using User.Application.Features.Orders;
+using LoginUserCommand = User.Application.Features.Orders.LoginUserCommand;
 
 namespace User.API.Controllers
 {
@@ -12,25 +12,22 @@ namespace User.API.Controllers
         private readonly IMediator _mediator = mediator;
 
         [HttpGet("{nickName}")]
-        public async Task<ActionResult<UserResponse>> GetUser([FromRoute] string nickName)
+        public async Task<ActionResult<UserResponse>> GetUser([FromRoute] GetUserQuery query)
         {
-            var query = new GetUserQuery(nickName);
             var result = await _mediator.Send((query));
             return Ok(result);
         }
 
         [HttpPost("register")]
-        public async Task<ActionResult<AuthResponse>> CreateUser([FromBody] RegisterDto registerDto)
+        public async Task<ActionResult<AuthResponse>> CreateUser([FromBody] CreateUserCommand command)
         {
-            var command = new CreateUserCommand(registerDto.NickName, registerDto.Email, registerDto.Password);
             var result = await _mediator.Send(command);
             return Ok(result);
         }
 
         [HttpPost("login")]
-        public async Task<ActionResult<AuthResponse>> LoginUser([FromBody] LoginDto loginDto)
+        public async Task<ActionResult<AuthResponse>> LoginUser([FromBody] LoginUserCommand command)
         {
-            var command = new LoginUserCommand(loginDto.Email, loginDto.Password);
             var result = await _mediator.Send(command);
             return Ok(result);
         }
