+++ Services/User/User.Infrastructure/Helpers/TokenManipulation.cs
