+++ Services/User/User.Infrastructure/User.Infrastructure.csproj
@@ -17,7 +17,7 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\Auth.Application\User.Application.csproj" />
+    <ProjectReference Include="..\User.Application\User.Application.csproj" />
   </ItemGroup>
 
 </Project>
